commands: move edit code to edit.rs
